requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
fasttext-wheel>=0.9.2
